Refactorin procuring entity
