+++ openprocurement/tender/cfaselectionua/models/submodels/ProcuringEntity.csv
@@ -1,0 +1,5 @@
+rolename,name_ru,additionalIdentifiers,kind,name,address,additionalContactPoints,name_en,identifier,contactPoint,__parent__
+edit_active.tendering,1,1,,1,1,1,1,1,1,
+default,1,1,1,1,1,1,1,1,1,
+embedded,1,1,1,1,1,1,1,1,1,
+view,1,1,1,1,1,1,1,1,1,
