+++ openprocurement/tender/cfaselectionua/models/tender.py
@@ -8,7 +8,7 @@ from openprocurement.tender.cfaselectionua.models.submodels.contract import Cont
 from openprocurement.tender.cfaselectionua.models.submodels.feature import Feature
 from openprocurement.tender.cfaselectionua.models.submodels.item import Item
 from openprocurement.tender.cfaselectionua.models.submodels.lot import Lot
-from openprocurement.tender.cfaselectionua.models.submodels.organization import ProcuringEntity
+from openprocurement.tender.cfaselectionua.models.submodels.organizationAndPocuringEntity import ProcuringEntity
 from schematics.types import StringType, IntType, URLType, BooleanType
 from schematics.types.compound import ModelType
 from zope.interface import implementer, provider
